Optimize fab file with CheckTry class. Check sphinx before building. Add mysql and postgress python libs check, flup and settings permisiions

+++ fabfile.py
@@ -2,12 +2,13 @@
 
 import os
 import shutil
+import stat
 import sys
 
 # START SETTINGS
 DOCS_OUT_DIR = "html"
 STATIC_DIR = "static"
-TMP_DIRECTORY = "tmp/limited"
+TMP_DIRECTORY = "tmp"
 TEST_SETTINGS = "settings"
 CHMOD = 644
 # END SETTINGS
@@ -16,10 +17,12 @@ CHMOD = 644
 PROJECT_PATH = os.path.realpath( os.path.dirname( __file__ ) )
 MANAGE_PATH = os.path.join( PROJECT_PATH, "manage.py" )
 STATIC_PATH = os.path.join( PROJECT_PATH, STATIC_DIR )
+TMP_PATH = os.path.join( PROJECT_PATH, TMP_DIRECTORY )
 HG_PATH = os.path.join( PROJECT_PATH, ".hg" )
 
 __all__ = ['check', 'clear', 'docs', 'static', 'test', 'update']
 
+# Helpers
 def list_files( path, ends=None ):
     """
     Return list with full path paths to all files under ``path`` directory.
@@ -41,6 +44,31 @@ def list_files( path, ends=None ):
     return _list_files( path, [] )
 
 
+class CheckTry:
+    """
+    run ``self.func`` in try..except block and check for ImportError, Exception, else.
+    At witch print ``self.import_out`` and ``self.except_out`` and ``self.success_out``
+    """
+
+    def __init__(self):
+        self.func = None
+        self.import_out = None
+        self.except_out = None
+        self.success_out = None
+
+    def check(self):
+        try:
+            self.func( )
+        except ImportError:
+            print( self.import_out )
+        except Exception:
+            print( self.except_out )
+        else:
+            print( self.success_out )
+
+
+# main commands
+#  also listed in ``__all__``
 def test( settings=None ):
     """
     Run limited tests.
@@ -52,6 +80,8 @@ def test( settings=None ):
     MANAGE_PATH = os.path.join( PROJECT_PATH, "manage.py" )
     RUN = "python {manage} test {app} --settings={settings}; return 0"
     clear( )
+    if not os.path.exists( TMP_PATH ):
+        os.mkdir( TMP_PATH )
     os.system( "clear" )
     os.system( RUN.format( manage=MANAGE_PATH, app="lviewer", settings=SETTINGS ) )
     os.system( RUN.format( manage=MANAGE_PATH, app="limited", settings=SETTINGS ) )
@@ -77,21 +107,25 @@ def static( settings=None ):
     files = list_files( STATIC, ("css", "js",) )
     for item in files:
         os.system( "gzip -6 -c {0} > {0}.gz".format( item ) )
-    os.system( "find {0} -type f -exec chmod {1} {{}} \;".format( STATIC_PATH, CHMOD  ) )
+    os.system( "find {0} -type f -exec chmod {1} {{}} \;".format( STATIC_PATH, CHMOD ) )
 
 
 def docs():
     """
     Build docs with sphinx
     """
-    print( "[ INFO ] Build documentation " )
-    DOCS_BUILD = "sphinx-build -b html -d {tmp} {source} {out}"
-    DOCS_OUT_PATH = os.path.join( PROJECT_PATH, DOCS_OUT_DIR )
-    DOCS_SOURCE_PATH = os.path.join( PROJECT_PATH, "docs" )
-    if os.path.exists( DOCS_OUT_PATH ):
-        shutil.rmtree( DOCS_OUT_PATH )
-    os.system( DOCS_BUILD.format( tmp=TMP_DIRECTORY, source=DOCS_SOURCE_PATH, out=DOCS_OUT_PATH ) )
-    os.system( "find {0} -type f -exec chmod {1} {{}} \;".format( DOCS_OUT_PATH, CHMOD ) )
+    if os.path.exists( "/usr/local/bin/sphinx-build" ) or os.path.exists( "/usr/bin/sphinx-build" ):
+        print( "[ INFO ] Build documentation " )
+        DOCS_BUILD = "sphinx-build -b html -d {tmp} {source} {out}"
+        DOCS_TMP = os.path.join( TMP_PATH, "docs" )
+        DOCS_OUT_PATH = os.path.join( PROJECT_PATH, DOCS_OUT_DIR )
+        DOCS_SOURCE_PATH = os.path.join( PROJECT_PATH, "docs" )
+        if os.path.exists( DOCS_OUT_PATH ):
+            shutil.rmtree( DOCS_OUT_PATH )
+        os.system( DOCS_BUILD.format( tmp=DOCS_TMP, source=DOCS_SOURCE_PATH, out=DOCS_OUT_PATH ) )
+        os.system( "find {0} -type f -exec chmod {1} {{}} \;".format( DOCS_OUT_PATH, CHMOD ) )
+    else:
+        print( "[ INFO ] No sphinx-build found" )
 
 
 def clear():
@@ -110,7 +144,6 @@ def check( tag="default" ):
     """
     Check installed programs, libraries and new changes in default branch
     """
-    RUN = "hg incoming --branch {branch} --template '{template}';return 0"
     print( "[ INFO ] Check installed programs and limited updates" )
     print( "" )
     # Easy install
@@ -121,28 +154,44 @@ def check( tag="default" ):
         print( "\t sudo apt-get install python-setuptools" )
 
     # Django
-    try:
-        from django import VERSION
-
-        if VERSION[1] < 3:
+    def _django():
+        if __import__( 'django' ).VERSION[1] < 3:
             raise
-    except ImportError:
-        print( "[ INFO ] No django exists" )
-        print( "\t sudo easy_install -U django" )
-    except Exception:
-        print( "[ INFO ] Django 1.3 version is needed" )
-        print( "\t sudo easy_install -U django" )
-    else:
-        print( "[ INFO ] Django exists" )
+
+    django = CheckTry( )
+    django.func = _django
+    django.import_out = "[ INFO ] No django exists\n\t sudo easy_install -U django"
+    django.except_out = "[ INFO ] Django 1.3 version is needed\n\t sudo easy_install -U django"
+    django.success_out = "[ INFO ] Django exists"
+    django.check( )
+
+    # flup
+    flup = CheckTry( )
+    flup.func = lambda: __import__( "flup" )
+    flup.import_out = "[ INFO ] No Python flup library exists\n\t sudo apt-get install python-flup"
+    flup.success_out = "[ INFO ] Python flup library exists"
+    flup.check( )
+
+    # MySQLdb
+    mysqldb = CheckTry( )
+    mysqldb.func = lambda: __import__( "MySQLdb" )
+    mysqldb.import_out = "[ INFO ] No Python MySQLdb library exists\n\t sudo apt-get install python-mysqldb"
+    mysqldb.success_out = "[ INFO ] Python MySQLdb library exists"
+    mysqldb.check( )
+
+    # Psycopg
+    psycopg = CheckTry( )
+    psycopg.func = lambda: __import__( "psycopg2" )
+    psycopg.import_out = "[ INFO ] No Python Psycopg library exists\n\t sudo apt-get install python-psycopg2"
+    psycopg.success_out = "[ INFO ] Python Psycopg library exists"
+    psycopg.check( )
 
     # PIL
-    try:
-        import PIL
-    except ImportError:
-        print( "[ INFO ] No Python image library exists" )
-        print( "\t sudo apt-get install python-imaging" )
-    else:
-        print( "[ INFO ] Python image library exists" )
+    pil = CheckTry( )
+    pil.func = lambda: __import__( "PIL" )
+    pil.import_out = "[ INFO ] No Python image library exists\n\t sudo apt-get install python-imaging"
+    pil.success_out = "[ INFO ] Python image library exists"
+    pil.check( )
 
     # Sphinx
     if os.path.exists( "/usr/local/bin/sphinx-build" ) or os.path.exists( "/usr/bin/sphinx-build" ):
@@ -152,11 +201,32 @@ def check( tag="default" ):
         print( "\t sudo apt-get install python-sphinx" )
         print( "\t sudo easy_install -U sphinx" )
 
+    #Settings file permissions
+    files = []
+    for item in os.listdir( PROJECT_PATH ):
+        full_path = os.path.join( PROJECT_PATH, item )
+        if os.path.isfile( full_path ) and 'setting' in item.lower( ):
+            mode = os.stat( full_path ).st_mode
+            # other
+            if bool( mode & stat.S_IROTH ) and bool( mode & stat.S_IROTH ):
+                files.append( item )
+    if files == []:
+        print( "[ INFO ] Settings file permissions right" )
+    else:
+        print( "[ INFO ] Settings file permissions not right" )
+        for file in files:
+            print( "\t Other can read file '{0}'".format( file ) )
+
     # HG updates
     print( "" )
+    RUN = "hg incoming --branch {branch} --template '{{rev}} {{desc|firstline}}\n'; return 0"
+    RUN2 = "hg log -l 5 --template '{{rev}} {{desc|firstline}}\n'"
     if os.path.exists( HG_PATH ):
         print( "[ INFO ] LimitedFM updates" )
-        os.system( RUN.format( branch=tag, template="{rev} {desc|firstline}" ) )
+        os.system( RUN.format( branch=tag ) )
+        print( "" )
+        print( "[ INFO ] LimitedFM hg log" )
+        os.system( RUN2.format( ) )
     else:
         print( "[ INFO ] No mercurial vcs exists, can't check for updates" )
